Fluff C2: split document_routes/common.py into schemas / rendering, move session helpers
Executes audit M2.3 (FLUFF_CUT_ORDER Phase C2): common.py (1,161 lines)
split by purpose —

- schemas.py (213): shared DTOs, contract constants, restored EditMode
- rendering.py (644): DOCX/PDF/HTML rendering + export-filename helpers
- edit_sessions.py: _active_session_cutoff/_active_edit_sessions moved
  to their only consumer
- common.py (331): namespace hub re-exporting schemas + rendering for
  the route groups' star imports, plus the cross-cutting DB/ownership
  helpers; zero caller churn

Restores `EditMode` (lost in the #171 router split), which had left
EditInstructionRequest an incomplete pydantic model — POST
/documents/{id}/edit-instructions would 500 at request time; invalid
modes now 422 at the boundary.

Audit M2.4 (core→api layering inversion in capabilities.py) verified
already resolved on master: `grep -rn "from app.api" backend/app/core`
returns nothing; the named symbols left with the modules surface.

Full backend suite: 3 failed (pre-existing macOS sandbox rlimit
baseline, test_phase3_sandbox), 806 passed, 11 skipped, 1 xfailed.
